requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.14
django-browser-reload==1.18.0
django-htmx==1.23.0
django-import-export==4.3.7
django-mssql-backend==2.8.1
django-tailwind==4.0.1
et_xmlfile==2.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mssql-django==1.5
numpy==2.2.5
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tabulate==0.9.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.5.0
xlrd==2.0.1
xlwt==1.3.0
